refactor * FILTER_MULTISELECT_MAX * DEFAULT_ORDER_TYPE * all filters to filters.ts
