+++ ts-client/src/dealExplorerClient/utils.ts
@@ -1,6 +1,10 @@
 import { ethers } from "ethers";
+import type {OrderType} from "./types/filters.js";
 
 export const DEFAULT_TOKEN_VALUE_ROUNDING = 3
+// Max to select per 1 multiselect filter.
+export const FILTER_MULTISELECT_MAX = 100;
+export const DEFAULT_ORDER_TYPE: OrderType = "desc";
 
 export function tokenValueToRounded(
   value: string | bigint,
